added database for quiz and quizcard, added stats screen nav and implemented the intent to create new quiz

+++ app/src/main/java/stelztech/youknowehv4/database/carddeck/CardDeckDao.java
@@ -193,7 +193,7 @@ public class CardDeckDao extends DbContentProvider implements ICardDeckDao, ICar
     }
 
     @Override
-    public long createCardDeck(int cardId, int deckId) {
+    public int createCardDeck(int cardId, int deckId) {
         ContentValues values = new ContentValues();
